Cleanup: Move form widget to forms.py
